- Updated all modules to inject actions via constructor - Removed middleware, using explicit policy checks - Ensured all business logic is handled by action classes and repositories - Fixed type errors and standardized DTO usage across modules

+++ Modules/Billing/DTOs/StripeDTO.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 readonly class StripeDTO
 {
     public function __construct(
-        public ?int $amount,
+        public ?float $amount,
         public ?string $currency,
         public ?string $source,
         public ?string $description
@@ -23,7 +23,7 @@ readonly class StripeDTO
     public static function fromArray(array $data): self
     {
         return new self(
-            isset($data['amount']) ? (int)$data['amount'] : null,
+            isset($data['amount']) ? (int) $data['amount'] : null,
             $data['currency'] ?? 'usd',
             $data['stripeToken'] ?? null,
             $data['description'] ?? 'KalimeroMK E-comm'
